chore: update auth code in reg to use basic v5
