--- bciers/apps/registration/app/api/auth/[...nextauth]/authOptions.ts
@@ -1,4 +1,0 @@
-import { NextAuthOptions } from "next-auth";
-import { authOptions as reg1AuthOptions } from "@/app/utils/auth/authOptions";
-
-export const authOptions = reg1AuthOptions as NextAuthOptions;
