+++ bciers/apps/registration/app/api/auth/token/route.ts
@@ -1,10 +1,12 @@
-import { getToken } from "next-auth/jwt";
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { getToken } from "@auth/core/jwt";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(request: NextRequest) {
+  // @ts-expect-error
   const token = await getToken({
     req: request,
-    secret: process.env.NEXTAUTH_SECRET,
+    secret: `${process.env.NEXTAUTH_SECRET}`,
   });
   return NextResponse.json(token, { status: 200 });
 }
