+++ bciers/apps/registration/app/components/operations/OperationsPage.tsx
@@ -3,8 +3,7 @@ import Link from "next/link";
 import { Button } from "@mui/material";
 import { Suspense } from "react";
 import Loading from "@bciers/components/loading/SkeletonGrid";
-import { getServerSession } from "next-auth";
-import { authOptions } from "../../api/auth/[...nextauth]/route";
+import { auth } from "../../../auth";
 import { OperationsSearchParams } from "@/app/components/operations/types";
 import Note, {
   registrationRequestNote,
@@ -19,9 +18,9 @@ export default async function OperationsPage({
 }) {
   // 👤 Use NextAuth.js hook to get information about the user's session
   /* When calling from the server-side i.e., in Route Handlers, React Server Components, API routes,
-   * getServerSession requires passing the same object you would pass to NextAuth
+   * next-auth uses a single method.
    */
-  const session = await getServerSession(authOptions);
+  const session = await auth();
   // @ts-ignore
   const role = session?.user?.app_role;
   const message = role?.includes("cas")
