+++ bciers/apps/registration/app/layout.tsx
@@ -7,14 +7,13 @@ You should not manually add <head> tags such as <title> and <meta> to root layou
 
 // eslint-disable-next-line import/extensions
 import "@bciers/styles/globals.css";
-import SessionProvider from "@bciers/components/auth/SessionProvider";
+import SessionProvider from "../auth/SessionProvider";
 import type { Metadata, Viewport } from "next";
 import Footer from "@bciers/components/layout/Footer";
 import Header from "@bciers/components/layout/Header";
 // 🏷 import {named} can be significantly slower than import default
 import Box from "@mui/material/Box";
-import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { auth } from "../auth";
 import { PublicEnvScript } from "next-runtime-env";
 import ThemeProvider from "@mui/material/styles/ThemeProvider";
 import { theme } from "@bciers/components";
@@ -42,7 +41,7 @@ export default async function RootLayout({
 }) {
   //🪝 Wrap the returned auth session in the "use client" version of NextAuth SessionProvider so to expose the useSession() hook in client components
   // Session properties come from client/app/api/auth/[...nextauth]/route.ts
-  const session = await getServerSession(authOptions);
+  const session = await auth();
 
   return (
     <html lang="en">
