+++ bciers/apps/registration1/app/utils/getUserFullName.ts
@@ -1,3 +1,4 @@
+//@ts-nocheck
 import { Session } from "next-auth";
 
 const getUserFullName = (session?: Session | null) => {
