+++ bciers/libs/shared/components/src/lib/form/components/MultiStepButtons.tsx
@@ -31,7 +31,7 @@ const MultiStepButtons: React.FunctionComponent<MultiStepButtonsProps> = ({
   const isDisabled = disabled || isSubmitting;
   const { data: session } = useSession();
 
-  const isIndustryUser = session?.user.app_role?.includes("industry");
+  const isIndustryUser = session?.user?.app_role?.includes("industry");
   return (
     <div className={`flex w-full mt-8 justify-between ${classNames}`}>
       {cancelUrl && (
