+++ bciers/libs/shared/components/src/lib/form/widgets/FileWidget.tsx
@@ -167,7 +167,7 @@ const FileWidget = ({
 
   const { data: session } = useSession();
   const isCasInternal =
-    session?.user.app_role?.includes("cas") &&
+    session?.user?.app_role?.includes("cas") &&
     !session?.user.app_role?.includes("pending");
 
   const hiddenFileInput = useRef() as MutableRefObject<HTMLInputElement>;
